requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.11.0+cpu
torchvision==0.12.0+cpu
timm==0.6.7
hydra-core==1.2.0